全局异常处理（1）

+++ src/main/java/com/shop/ssm/utils/Constant.java
@@ -4,8 +4,8 @@ package com.shop.ssm.utils;
  * Created by Administrator on 2017/9/26.
  */
 public final class Constant {
-    public final static short IS_EXIST=0;//存在，未删除
-    public final static short IS_DELETE=1;//不存在，删除
+    public final static Integer IS_EXIST=0;//存在，未删除
+    public final static Integer IS_DELETE=1;//不存在，删除
     public final static String SUCCESS="200";//成功
     public final static String FALSE="400";//失败
